% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel_prob.R
\name{mendel_prob}
\alias{mendel_prob}
\title{mendel_prob}
\usage{
mendel_prob(num_family, num_case, gene_freq, min_num_total_mutation,
  min_num_family_mutation)
}
\arguments{
\item{num_family}{number of families}

\item{num_case}{number of unrelated cases}

\item{gene_freq}{proportion of affected individuals explained by one gene}

\item{min_num_total_mutation}{minimum number of families and cases with pathogenic variant(s) in the same gene}

\item{min_num_family_mutation}{minimum number of families with pathogenic variant(s) in the same gene}
}
\value{
probability
}
\description{
Determine the probability of detecting a minimum number of families or unrelated cases with pathogenic variants in the same gene.
}
\examples{
mendel_prob(125, 500, 0.005, 2, 1)
}
