% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel_prob.R
\name{mendel_prob_with_family_ratio}
\alias{mendel_prob_with_family_ratio}
\title{mendel_prob_with_family_ratio}
\usage{
mendel_prob_with_family_ratio(num_probands_total, proband_prop, gene_freq,
  min_num_variants, min_num_probands_variants)
}
\description{
internal use only.
}
