% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel_prob.R
\name{mendel_sample_size}
\alias{mendel_sample_size}
\title{mendel_sample_size}
\usage{
mendel_sample_size(prob = 0.8, proband_prop = NULL, gene_freq = 0.005,
  min_num_variants = 2, min_num_probands_variants = 0)
}
\arguments{
\item{prob}{probability}

\item{proband_prop}{proportion of proband being a member of a family, NULL if the proportion is not specified}

\item{gene_freq}{proportion of affected individuals explained by one gene}

\item{min_num_variants}{minimum number of probands with pathogenic variant(s) in the same gene}

\item{min_num_probands_variants}{minimum number of proband being a member of a family with pathogenic variant(s) in the same gene (cannot be bigger than min_num_variants)}
}
\value{
return number of samples (num_family + num_case) needed to reach desired probability
}
\description{
Determine the number of families/cases which need to be screened to detect a minimum number of observations of potentially pathogenic variants within the same gene.
}
\examples{
mendel_sample_size(0.8, 0.2, 0.005, 2, 1)
}
