% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel_prob.R
\name{mendel_prob_with_family_ratio}
\alias{mendel_prob_with_family_ratio}
\title{mendel_prob_with_family_ratio}
\usage{
mendel_prob_with_family_ratio(num_family_plus_case, family_prop, gene_freq,
  min_num_total_mutation, min_num_family_mutation)
}
\description{
internal use only.
}
