% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel_prob.R
\name{mendel_sample_size}
\alias{mendel_sample_size}
\title{mendel_sample_size}
\usage{
mendel_sample_size(prob, family_prop, gene_freq, min_num_total_mutation,
  min_num_family_mutation)
}
\arguments{
\item{prob}{probability}

\item{family_prop}{proportion of family}

\item{gene_freq}{proportion of affected individuals explained by one gene}

\item{min_num_total_mutation}{minimum number of families and cases with pathogenic variant(s) in the same gene}

\item{min_num_family_mutation}{minimum number of families with pathogenic variant(s) in the same gene}
}
\value{
(num_family, num_case) needed to reach desired probability
}
\description{
Determine the number of families/cases which need to be screened to detect a minimum number of observations of potentially pathogenic variants within the same gene.
}
\examples{
mendel_sample_size(0.8, 0.2, 0.005, 2, 1)
}
